requirements.txt
Flask==3.1.0
Flask-Cors==3.1.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
SpeechRecognition==3.8.1
librosa==0.10.1
numpy==2.1.3
nltk==3.8.1
python-Levenshtein==0.21.0
